requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
asyncpg==0.29.0
base58==2.1.1
bcrypt==4.1.2
bitcoin-utils-fork-minimal==0.4.11.6
blacksheep==2.0.7
block-io==2.0.6
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.1.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.7
ecdsa==0.18.0
essentials==1.1.5
essentials-openapi==1.0.9
greenlet==3.0.3
guardpost==1.0.2
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
itsdangerous==2.1.2
Mako==1.3.3
MarkupSafe==2.1.3
marshmallow==3.21.2
mpmath==1.3.0
packaging==24.0
pluggy==1.5.0
psycopg2==2.9.9
pycparser==2.21
pycryptodome==3.20.0
pydantic==2.5.2
pydantic-settings==2.2.1
pydantic_core==2.14.5
PyJWT==2.6.0
pytest==8.3.2
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.31.0
rodi==2.0.6
setuptools==69.5.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.28
sqlmodel==0.0.16
sympy==1.12
typing_extensions==4.10.0
urllib3==2.2.1
uvicorn==0.22.0
websockets==10.4
